test : 원시값 포장으로 인한 테스트 코드 분리 & 리팩토링
